bfq-iosched: fix compile warning about possibly used uninitialized
A local variable is declared and the use case is clear that the
address of it is passed into a function.  However gcc isn't clever
enough to see this as setting the variable.

We have a macro in compiler.h for handling this that doesn't actually
add code or waste cycles by doing a pointless init.  The file does
use it in other places already; we extend that to this location.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>