armv8: aarch32: Add freescale_aarch32.config for AArch32 support
This patch is to add freescale_aarch32.config for AArch32 support on
ARMv8 platforms.

PCI_KERYSTONE is enabled in multi_v7_defconfig. As there is some
compling erros for it in this kernel version, this patch is to disable
it as a workaround. In the latest upstream kernel, there is not such
problem.

Signed-off-by: Alison Wang <alison.wang@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>