ARM64: zynqmp: Block on set requirement when powering on a domain
commit  66c15e5b9e1c60c64636e5a43a3ccc56c78e53b5 from
https://github.com/Xilinx/linux-xlnx.git

If the zynqmp_pm_set_requirement call is non-blocking, the return value
represents only the status of sending IPI message to PMU (status 0 means
that the message arguments and the IPI are sent to the PMU). Consequently,
when the zynqmp_pm_set_requirement call returns the PMU may not have
completed the processing of the call. In that case, if the device driver
attempts to access a domain which is still off it would crash. Moreover,
the result of actual processing of the call is not returned, so a power
on failure cannot be captured.
Make the call blocking, so the returned value represents the status
of processing the zynqmp_pm_set_requirement call.

Signed-off-by: Mirela Simonovic <mirela.simonovic@aggios.com>
Acked-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>