misc: sram: Introduce 'protect-exec' reserved sram area type
Introduce a new 'protect-exec' reserved sram area type which is
functionally the same as the existing 'pool' sram region type with the
added requirement that it be maintained as read-only and executable
while allowing for an arbitrary number of drivers to share the space.

Some platforms, like many ARM SoCs, require to be able to run code from
on-chip memory like SRAM for tasks like reconfiguring the SDRAM
controller or entering low-power sleep modes. This introduces a common
way to maintain a region of sram as read-only and executable and also
introduces a helper function, sram_exec_copy, which allows for copying
data to this protected region while maintaining locking to avoid conflicts
between multiple users of the same space.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
(cherry picked from commit 656c860c6c8a8851becde1818aa4130858efaac3)