ARM: OMAP2+: sleep33/43xx: Make sleep actions configurable
Add an argument to the sleep33xx and sleep43xx code to allow us to set
flags to determine which portions of the code get called in order to use
the same code for cpuidle sequence as well. This patch allows us to
decide whether or not we flush and disable caches, save EMIF context,
put the memory into self refresh and disable the EMIF, and/or invoke
the wkup_m3 when entering into WFI.

Also for sleep43xx, clean up the WFI_FLAG_RTC_ONLY checking to match the
flags introduced here.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
(cherry picked from commit e876de4c2c760087089af3086ee7810e39a50430)