staging: fsl-dpaa2: Implement ndo_select_queue
Normally skb_record_rx_queue would be enough to guarantee flow affinity
through the networking stack on forwarding scenarios. However, for the
termination scenarios, on the sender side, the initial flow affinity is
determined based on the skb_hash, which is expected to hit only a few
Tx queues when there is a small number of flows.

For this final scenario, send the frame on the Tx queue associated with
the CPU that does the kernel space processing of the Tx frame. This
ensures that the confirmation frame is processed on the same CPU that
did the enqueue in the first place, since there's a 1:1 mapping between
Tx and Tx conf queues, and the Tx conf queues are CPU affine.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>