Revert "rtmutex/rt: don't BUG for -EDEADLK when detect_deadlock is off"
This reverts commit baf9cf54e6b52ae7c9661e869719251e87225efd.

The fix was designed for early 3.x-rt releases, and even then for
a fixed subset, since the issue was resolved in later stable updates
in a different way.

While having it present should not present any runtime issues, there
is no point in deviating from the main -rt release with this.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>