dmaengine: xilinx_dma: Check for channel idle state before submitting dma descriptor
commit  895ba71c3c96ca5b9691c520f493e9976d4df1d5 from
https://github.com/Xilinx/linux-xlnx.git

Add variable for checking channel idle state to ensure that dma descriptor is not
Submitted when DMA engine is in progress.

This will avoids the pollling for a bit in the status register to know
Dma state in the driver hot path.

Reviewed-by: Jose Abreu <joabreu@synopsys.com>
Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>