pwm: lpss: Avoid reconfiguring while UPDATE bit is still enabled
commit 151d1093ccf71e3d031d63ed2c06028a7cf1c401 from
git://git@github.com:01org/linux-intel-4.9.git

PWM Configuration register has SW_UPDATE bit that is set when a new
configuration is written to the register. The bit is automatically
cleared at the start of the next output cycle by the IP block.

If one writes a new configuration to the register while it still has
the bit enabled, PWM may freeze. That is, while one can still write
to the register, it won't have an effect. Thus, we try to sleep long
enough that the bit gets cleared and make sure the bit is not
enabled while we update the configuration.

Upstream status:
Patch will be submitted to linux pwm mailing list January 2017.

Change-Id: Ifef1c3c7845b5b6e019d500977800803099c6974
Signed-off-by: Ilkka Koskinen <ilkka.koskinen@intel.com>
Signed-off-by: Richard A. Griffiths <richard.a.griffiths@intel.com>