watchdog: Introduce watchdog_stop_on_unregister helper
commit bb292ac1c6028344013309a309b44dc691581825 upstream.

Many watchdog drivers explicitly stop the watchdog when unregistering it.
While it is unclear if this is actually needed (the whatdog should not be
running at that time if it can be stopped), introduce a helper to
explicitly stop the watchdog in the watchdog core when unregistering it.
This helps reducing driver code size while retaining functionality.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Liwei Song <liwei.song@windriver.com>