ASoC: Intel: Skylake: Check for memory reclaim bit
commit 1ec405c1ab971ec7aef632ddfbef9a3ccf9247d1 from
git://git@github.com:01org/linux-intel-4.9.git

Memory reclaim bit indicates whether legacy DMA memory
is managed by FW. It would be set by firmware if DMA
is to be used. Check for the memory reclaimed bit during
the probe. If the bit is not set then fail the probe.

Change-Id: I8a926ea2b4f86b1f7d66749d2e9809f23308c51c
Signed-off-by: Leoni Prodduvaka <leoni.prodduvaka@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/13482
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Nc, Shreyas <shreyas.nc@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>