remoteproc: add infrastructure to support user-space loading/booting
The remoteproc infrastructure is enhanced to allow remoteproc drivers
to support userspace driven loading and booting mechanism. This is
done in preparation to allow the TI Keystone remoteproc driver to
support a userspace based loader.

A new field 'use_userspace_loader' is added to the rproc structure,
and this field is expected to be set alongside configuring the
'auto_boot' flag to false by remoteproc drivers wishing to bypass
the in-kernel loading and boot mechanism. The remoteproc core skips
looking for firmware or loading any firmware segments using these
two state flags.

The remoteproc drivers are expected to invoke rproc_boot() and
rproc_shutdown() for triggering the boot and shutdown of the remote
processor after the loading is completed and the resource table
information is published to the remoteproc driver. The resource
table is processed in-line during the rproc_boot() invocation.

The interface and implementation details to userspace is left to
the individual remoteproc drivers.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sam Nelson <sam.nelson@ti.com>
(cherry picked from commit 1c15e38f32b1ab1fd11a14ca8670dcfa56651f13)