mxc IPUv3: common: avoid misoperation on the global string array
According to the C reference manual, whether identical string literals are
distinct is implementation defined, and the behavior of a program that attempts
to alter a string literal is undefined. And it will result an unpredictable error.
So change the pointer of string array instead of alterring a string literal.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>