mtd: spi-nor: Fix incorrect status register update
commit  7b994e66ab4ff7ea5d7bb12caf21a3c3b1d0650b from
https://github.com/Xilinx/linux-xlnx.git

To check whether flash is ready or not, we use status register
command (SR). so for QSPI dual prallel configuration, we have to
wait for ready status of both flash devices. we are doing that,
but due to the below & operation, we are sending incorrect status
update to layer, causing invalid data write/read to/from the devices.
so this patch fixes that.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>