net: macb: SW workaround for link instability on DP83867
commit  7557928cecc63c43191eb23b41f66ad93c497f1b from
https://github.com/Xilinx/linux-xlnx.git

RX_CTRL of DP83867 has to be strapped to MODE 3 or 4 according to
TI datasheet. If this is not done link instability may be observed
with some partners. This was observed on ZCU102 at 100Mbps and
also previously with on SGMII DC when connected to switches.
From communication with TI, we learnt that the following can be
used as a SW workaround to solve these issues.
Added this workaround with a DT property to make it configurable.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>