misc: vcu: updated error handling.
commit  75334bd733b324c967ebecb70e655cab8e09996c from
https://github.com/Xilinx/linux-xlnx.git

pll_ref clocksource can be a fixed or si570_1 type.
If there is a fixed clock in the pll_ref, we cann't set the
clock in that pll_ref clock source. In this case, we should
bypass the error with the warning only. this will work for
any kind of pll_ref clocksource.

Signed-off-by: Dhaval Shah <dshah@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>