arm64: zynqmp: pm: Notify PM firmware that initialization is completed
commit  223b01f70fe419b30d5bfad7a7a1349a83e8f678 from
https://github.com/Xilinx/linux-xlnx.git

PM firmware requires notification when power management is enabled and
when the initialization is completed in order to finalize
system initialization (e.g. power down slaves that will not be used).

Implement PM_INIT_FINALIZE PM API, which is used to inform the PM
firmware that the initialization is completed. This call is made only
if CONFIG_PM is enabled in late_initcall_sync phase of the boot procedure.

Signed-off-by: Filip Drazic <filip.drazic@aggios.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>