drivers: net: ethernet: TSN-switch driver initial check-in
commit c60aebb39a7957861f34692283e380468f18b329 from
https://github.com/Xilinx/linux-xlnx.git

The Xilinx TSN Switch is an integrated solution provided with
3-ports, two ports connecting to network through MAC interfaces
and one port connecting to an internal End Point.
TSN switch driver allows processor to control and program the
TSN Switch hardware settings.
This misc driver will be used to provide the APIs for accessing Switch
Core resgisters. Most of the APIs are exposed to userspace by ioctl.

Signed-off-by: Saurabh Sengar <saurabhs@xilinx.com>
Acked-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>