remoteproc/omap: check for all timer ops for watchdog timers
The timers and watchdog timers are both optional for OMAP remoteproc
devices. The timers are rarely not defined though especially given
that they are a absolute must for any active power management. But,
there can potentially be a configuration where timers are not defined
and only watchdog timers are defined. All of these also require all
the timer ops to be defined, so ensure that all timer ops are checked
even for the watchdog timers, and be agnostic of the presence of
regular timers.

Fixes: 398afae34f11 ("remoteproc/omap: add watchdog functionality for remote processors")
Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit de8e0baeaefbb26b6a6276de36bf89aeb73c086d)