zynqmp: devicetree: Enabled CCI support for USB
commit  4c3cb2d4bd16aecd6897d2061e23ac22bf7dff20 from
https://github.com/Xilinx/linux-xlnx.git

This patch adds CCI support for USB when CCI is enabled in design.
This patch also adds 'reg' property for Xilinx USB 3.0 IP. The 'reg'
property is added in order to modify a register in that to enable
coherency in Hardware.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>