cpufreq: intel_pstate: Drop struct cpu_defaults
commit 2f49afc2a6f6ec4bacb2e50d5482ecc111b41ab5 upstream

The cpu_defaults structure is redundant, because it only contains
one member of type struct pstate_funcs which can be used directly
instead of struct cpu_defaults.

For this reason, drop struct cpu_defaults, use struct pstate_funcs
directly instead of it where applicable and rename all of the
variables of that type accordingly.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>