usb: dwc3: of-simple: Enable CCI support in dwc3 driver
commit  554c94c1bc51217fc3d95a4db305b3a1912c221b from
https://github.com/Xilinx/linux-xlnx.git

This patch adds CCI support in DWC3 driver when CCI is enabled in the
design. There are couple of registers which need to be modified for
making CCI enabled transfers in USB. This patch add the support for the
same.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>