soc: ti: add helper functions for use across knav netcp drivers
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

The helper functions that formats the descriptors and access the common
qmss queue related tasks are usable across all of knav netcp drivers.
Some of these such as crypto, srio etc are outside network driver. So
move these into knav_helpers under include/linux/soc/ti so that it can
be included in all of the knav netcp drivers.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>