drm: xilinx: dp: Add a module param for aux timeout
commit  de743f565e82ffea81c5989e2178643ad0418557 from
https://github.com/Xilinx/linux-xlnx.git

From testing with multiple monitors, the timeout value varies.
The spec describes the 25 msec, but some monitor requires a few
seconds wait time. So, the default timeout is set to 50 msec,
but make it configurable so that the value can be tuned.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>