ARM: dts: keystone-k2l: Add PSC reset control data
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

commit cb11101583f8fdc044cfcaa231ac76c1b3558448 ti-sdks

The Power Sleep Controller (PSC) module contains specific
memory-mapped registers that can be used to perform reset
management for the DSPs available on the SoC using specific
bits. The PSC is defined using a syscon node, and the reset
functionality is defined using a child syscon reset controller
node. Add the reset control data to this syscon reset controller
node for the Keystone 2 Lamarr SoC which the 4 different consumer
DSP devices can refer to for managing their resets.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>