Revert "drm/i2c: adv7511: Convert to drm_bridge"
commit  a752af6f0eae478caaed04e08b1cb34c48b0f393 from
https://github.com/Xilinx/linux-xlnx.git

This reverts commit e12c2f645557708932e15afdf77f5965549cf156
("drm/i2c: adv7511: Convert to drm_bridge").

This is a temporary solution to enable adv7511 based on the encoder
slave. The DRM driver needs to support the atomic modesetting
as well as the drm bridge. When those changes are in, this can be
removed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>