ASoC: Intel: board: Move cnl_rt274 clock setting to supply widget
commit 8f374fa96e84a60163306f25dc51926b104ab9c0 from
git://git@github.com:01org/linux-intel-4.9.git

During BE-BE loop, codec clocks were not set as it was a part of dai link ops
hw_params and no sound was heard due to this reason when use cases involve
BE-BE loop.

So, move codec clock setting as a part of supply widget and define routes
appropriately. Also use macro to define BE rate fixup and use it for both
fixup as well as clock computation.

Change-Id: Id5a08d2bd6024a61b601dbbe70ad99a52149da5e
Signed-off-by: Guneshwor Singh <guneshwor.o.singh@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/14595
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Tested-by: Avati, Santosh Kumar <santosh.kumar.avati@intel.com>