remoteproc: add api for retrieving the rproc alias id
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

commit 7e8e1d1906bce77106bb891588b5334d3970b25d ti-sdks

A new api, rproc_get_alias_id, is added to allow remoteproc/rpmsg
clients to be able to retrieve a fixed id associated with a rproc
device. Each of the rprocs is given an ida index as well, but this
is dynamic depending on the order rproc_alloc is called (probe order),
and also depends on what remote processor devices are instantiated.
The remoteproc alias id is therefore chosen, with the assumption that
the remoteproc devices are created with known alias ids in DT.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>