soc: ti: knav_qmss_queue: fix Linking RAM 0 Size register configuration
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

commit 05d0982c4b78302ff976c753ffaed3d3ecc17d13 ti-sdks

This patch corrects a configuration error in the Linking RAM 0 Size
register. Per spec, descriptors with index less than equal to
regions0_size is in region 0, in the case of NSS found in K2E/L/HK
where as for NSS lite found in K2G, descriptors with index less
than regions0_size is in region 0. So update the size accordingly
in the regions0_size bits of the linking ram size 0 register.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>