clocksource: timer-keystone: Add platform probe support to existing code
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

commit 4e24eba442464cdf8d9517e8e12587b9904aa939 ti-sdks

This patch fixes the case for k2g where in clocks are not
available at the time of keystone_timer_init function call.
K2HK/K2L/K2E devices are left untouched and the old way of
keystone_timer_init is preserved since it is init order sensitive.

On top of the existing code a platform driver is introduced so that
if clock is not yet ready probe can be deferred. This enables hres
timers for k2g.

Signed-off-by: Keerthy <j-keerthy@ti.com>
[t-kristo@ti.com: squashed a fix for section mismatch]
Signed-off-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>