clk: zynqmp: pll: Do correct rounding for max and min requirements
commit  f7f60245e1535f6b54e7a189aae84267c91cda18 from
https://github.com/Xilinx/linux-xlnx.git

The fbdiv for max value should be rounded down to make sure the rate
never goes above the max rate. On the other hand, the fbdiv for
the min should be rounded up so that the resulted rate is above
the minimum rate.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>